requirements.txt
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
hupper==1.12.1
idna==3.10
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-telegram-bot==22.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
uritemplate==4.1.1
urllib3==2.4.0
flask==2.2.3
python-telegram-bot==20.1
gspread==5.0.0
google-auth==2.14.1

